Refactored to allow testing of non-PBI Fabric items
